feat: CAN 멀티프레임(MULTI=1) — 8바이트를 넘는 페이로드 분할·재조립
CAN ID의 MULTI 비트는 자리만 있고 사용 규격도 구현도 없어서, CAN 유닛에서는 8바이트를
넘는 명령을 아예 보낼 수 없었다. 궤적 적재(C2)와 TimeSync 왕복 회신(19B)이 여기에
막혀 있었다. 확장 규격 §5를 펌웨어와 앱 양쪽에 구현한다.

  FIRST : [hdr][전체길이 BE16][데이터 5B]
  이후   : [hdr][데이터 7B]
  hdr    : bit7=FIRST, bit6=LAST, bit5:0=순번(0~63)
  최대   : 5 + 63×7 = 446바이트

송신은 호출부가 분할을 신경 쓰지 않는다 — can_transport_send()가 8바이트를 넘으면
자동으로 나눈다. 이에 맞춰 len 인자를 uint8_t → uint16_t로 넓혔고, 8바이트에서
잘라 버리던 can_reply()도 전체 페이로드를 보내도록 고쳤다.

수신은 (SrcId, OpCode)별 슬롯 4개로 재조립한다. FIRST 누락·순번 불일치·200ms
타임아웃이면 통째로 버린다 — 부분 실행은 절대 하지 않는다. 슬롯이 모자라면 가장
오래된 것을 회수한다(어차피 타임아웃 대상).

앱(rcube/can.py)에도 같은 규격으로 split_multiframe()과 Reassembler를 넣었다.
python-can 송신은 세그먼트를 순서대로 흘리고, 수신 루프는 MULTI=1이면 재조립이
끝난 뒤에만 상위로 올린다.

검증(파이썬 전수, 8항목): 길이 9~446 왕복 일치, 세그먼트 8B 이하, FIRST/LAST 플래그와
순번, 최대 길이 초과 거부, FIRST 누락 폐기, 순번 불일치 폐기, 타임아웃 폐기,
(src,op) 동시 재조립 분리.

이 테스트가 문서 오류를 하나 잡았다. §5에 "웨이포인트 37개"라고 적었는데 C2 헤더
4바이트를 빼먹은 계산이었다(4 + 37×12 = 448 > 446). 실제 상한은 36개이고 문서를
고쳤다.

+++ esp32/main/can_transport.h
@@ -26,9 +26,11 @@
 /* CAN 전송 계층 초기화 + 태스크 기동. node_id/term_id는 rcube_config 값. */
 esp_err_t can_transport_init(uint8_t node_id, uint8_t term_id);
 
-/* 임의 CAN 프레임 송신(저수준). 성공 ESP_OK. */
+/* 임의 CAN 프레임 송신(저수준). 성공 ESP_OK.
+ * len이 8을 넘으면 확장 규격 §5 멀티프레임(MULTI=1)으로 자동 분할한다
+ * (최대 RCUBE_CAN_REASSEMBLY_MAX=446바이트). 호출부는 분할을 신경 쓰지 않는다. */
 esp_err_t can_transport_send(uint8_t priority, uint8_t op_code,
-                             uint8_t dst, const uint8_t *data, uint8_t len);
+                             uint8_t dst, const uint8_t *data, uint16_t len);
 
 /* edge central(ECF=1)이 CAN 멤버를 기다리기 시작한다(기획서 7.4-4 [CMF=1 멤버]).
  *
